SistemaDeContabilidadGeneral/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-environ==0.11.2
django-filter==23.3
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
